Etc
+ Switched to a universal storyboard
+ Updated project settings
+ Fixed Xcode warnings
+ Tidied up

+++ LAN Scan/ScanLAN.m
@@ -46,7 +46,7 @@
     NSArray *b = [self.netMask componentsSeparatedByString:@"."];
     if ([self isIpAddressValid:self.localAddress] && (a.count == 4) && (b.count == 4)) {
         for (int i = 0; i<4; i++) {
-            int and = [[a objectAtIndex:i] integerValue] & [[b objectAtIndex:i] integerValue];
+            int and = (int)([[a objectAtIndex:i] integerValue] & [[b objectAtIndex:i] integerValue]);
             if (!self.baseAddress.length) {
                 self.baseAddress = [NSString stringWithFormat:@"%d", and];
             }
@@ -67,7 +67,7 @@
 
 - (void)pingAddress{
     self.currentHostAddress++;
-    NSString *address = [NSString stringWithFormat:@"%@%d", self.baseAddress, self.currentHostAddress];
+    NSString *address = [NSString stringWithFormat:@"%@%ld", self.baseAddress, (long)self.currentHostAddress];
     [SimplePingHelper ping:address target:self sel:@selector(pingResult:)];
     if (self.currentHostAddress>=254) {
         [self.timer invalidate];
@@ -82,8 +82,8 @@
     self.timerIterationNumber++;
     if (success.boolValue) {
         NSLog(@"SUCCESS");
-        NSString *deviceIPAddress = [[[[NSString stringWithFormat:@"%@%d", self.baseAddress, self.currentHostAddress] stringByReplacingOccurrencesOfString:@".0" withString:@"."] stringByReplacingOccurrencesOfString:@".00" withString:@"."] stringByReplacingOccurrencesOfString:@".." withString:@".0."];
-        NSString *deviceName = [self getHostFromIPAddress:[[NSString stringWithFormat:@"%@%d", self.baseAddress, self.currentHostAddress] cStringUsingEncoding:NSASCIIStringEncoding]];
+        NSString *deviceIPAddress = [[[[NSString stringWithFormat:@"%@%ld", self.baseAddress, (long)self.currentHostAddress] stringByReplacingOccurrencesOfString:@".0" withString:@"."] stringByReplacingOccurrencesOfString:@".00" withString:@"."] stringByReplacingOccurrencesOfString:@".." withString:@".0."];
+        NSString *deviceName = [self getHostFromIPAddress:[[NSString stringWithFormat:@"%@%ld", self.baseAddress, (long)self.currentHostAddress] cStringUsingEncoding:NSASCIIStringEncoding]];
         [self.delegate scanLANDidFindNewAdrress:deviceIPAddress havingHostName:deviceName];
     }
     else {
